chile-mineria-pbi/requirements.txt
pandas>=2.1.0
sqlalchemy>=2.0.0
openpyxl>=3.1.0
PyYAML>=6.0.1
pdfplumber>=0.11.0
pymysql>=1.1.1
